This is the first commit. These are the files used in the analysis.
These files contain the data used for the analysis for the paper submission for
Big Data & Society journal. "Examining political mobilization of online
communities through e-petitioning behavior in WeThePeople". They are here to support the reproducibility of this research.

e Please enter the commit message for your changes. Lines starting